Split subcommand.go into separate files to keep it a bit cleaner. Removed ctx from RunCommand because it's already captured in the closure of the command creation functions
